refactor(infrastructure): update package
