+++ transport/consumer/handler/setup.go
@@ -3,8 +3,8 @@ package consumerHandler
 import (
 	"context"
 
+	"github.com/alhamsya/boilerplate-go/infrastructure/middlewares/consumer"
 	"github.com/alhamsya/boilerplate-go/lib/managers/custom_log"
-	"github.com/alhamsya/boilerplate-go/middleware/consumer"
 	"github.com/alhamsya/boilerplate-go/transport/consumer/routers"
 )
