+++ transport/cron/handler/types.go
@@ -1,8 +1,8 @@
 package cronHandler
 
 import (
+	"github.com/alhamsya/boilerplate-go/infrastructure/middlewares/cron"
 	"github.com/alhamsya/boilerplate-go/lib/managers/config"
-	"github.com/alhamsya/boilerplate-go/middleware/cron"
 	"github.com/alhamsya/boilerplate-go/transport/cron/routers"
 	"github.com/robfig/cron/v3"
 )
